fix: resolve 13 bugs (2 P0, 5 P1, 6 P2) found in backend testing
P0 fixes:
- BUG-001: admin.py device lookup used Device.id instead of Device.device_id
- BUG-002: public.py weather endpoint crashed on API failure (added try-except → 503)

P1 fixes:
- BUG-003: replaced deprecated datetime.utcnow() with datetime.now(timezone.utc)
- BUG-004: added physical ForeignKey constraints for sensor_counts/camera_images
- BUG-006: added pyproject.toml with pytest-asyncio auto mode config
- BUG-007: hourly_aggregation kept UUID objects instead of converting to str

P2 fixes:
- BUG-008: removed duplicate httpx in requirements.txt
- BUG-009: separated heavy AI deps into requirements-ai.txt
- BUG-010: added missing DeviceStatusLog import in seed_data.py
- BUG-011: fixed weather_grade precip type hint (int → float)
- BUG-012: added localhost:5173 to default CORS origins
- BUG-013: added security note for JWT_SECRET_KEY in .env.example

+++ backend/app/services/weather.py
@@ -25,7 +25,7 @@ _cache: dict = {}
 CACHE_TTL = 3600  # 1 hour
 
 
-def _weather_grade(wmo: int, wind: float, precip: int) -> str:
+def _weather_grade(wmo: int, wind: float, precip: float) -> str:
     """Determine weather grade A/B/C per design doc Tab1 logic."""
     if wmo >= 95 or wind >= 15:
         return "C"
